+++ pluslib/HTTP/Controller.php
@@ -1,0 +1,6 @@
+<?php
+namespace pluslib\HTTP\Controller;
+
+abstract class Controller
+{
+}